+++ .github/workflows/ui-tests.yaml
@@ -47,7 +47,7 @@ on:
         required: false
         default: ""
   schedule:
-    - cron: "0 2 * * *"
+    - cron: "27 1 * * *"
 
 jobs:
   build-and-test:
